Adds Issues section with nvim-qt issue
nvim-qt does not display the cursor in terminal mode correctly. This
seems to have be an issue where the terminal in nvim reverses the
colours of the cursor. nvim-qt applies this as well and the color of the
cursor becomes black and almost invisible against the almost black
background.

The fix for this, which cannot be applied in the colour scheme, is to
set `hi TermCursor gui=reverse cterm=reverse` which then colours the
cursor correctly

Fixes #14